Mech Final - Autonomous Robot
This project was the final project to the UCSC Mechatronics course. It
involved building a fully autonomous robot with sensors we also built.
The robot was to find an ammo plunger, illuminated by track wires, and
depress the plunger to pick up balls. The robot would the navigate to a
target illuminated by a beacon emitting a certain frequency and drop off
the balls at two different target locations. The robot was to stay on
the arena, bordered by black tape.

The ES Framework files were provided, but editted by me. For some other
files, a template was provided by the instructors of the course (Gabe
Elkaim and Max Dunne) but completely restructured or filled with code I
wrote. Their names on some of the files denote that they created the
template, but the material inside was of my doing.